+++ src/components/Cocktail/CocktailIndex.vue
@@ -185,6 +185,7 @@ export default {
                     { name: this.$t('shelf.cocktails'), active: false, id: 'on_shelf' },
                     { name: this.$t('my-favorites'), active: false, id: 'favorites' },
                     { name: this.$t('cocktail.shared'), active: false, id: 'is_public' },
+                    { name: this.$t('shelf.locked_cocktails'), active: false, id: 'locked_user_cocktails' },
                 ],
                 abv: [
                     { name: this.$t('non-alcoholic'), min: null, max: 2, id: 'abv_non_alcoholic' },
@@ -479,6 +480,7 @@ export default {
             this.activeFilters.on_shelf = state.filter && state.filter.on_shelf ? state.filter.on_shelf : null
             this.activeFilters.bar_shelf = state.filter && state.filter.bar_shelf ? state.filter.bar_shelf : null
             this.activeFilters.locked_bar_cocktails = state.filter && state.filter.locked_bar_cocktails ? state.filter.locked_bar_cocktails : null
+            this.activeFilters.locked_user_cocktails = state.filter && state.filter.locked_user_cocktails ? state.filter.locked_user_cocktails : null
             this.activeFilters.favorites = state.filter && state.filter.favorites ? state.filter.favorites : null
             this.activeFilters.is_public = state.filter && state.filter.is_public ? state.filter.is_public : null
             this.activeFilters.total_ingredients = state.filter && state.filter.total_ingredients ? state.filter.total_ingredients : null
@@ -520,6 +522,7 @@ export default {
                 on_shelf: this.activeFilters.on_shelf,
                 bar_shelf: this.activeFilters.bar_shelf,
                 locked_bar_cocktails: this.activeFilters.locked_bar_cocktails,
+                locked_user_cocktails: this.activeFilters.locked_user_cocktails,
                 favorites: this.activeFilters.favorites,
                 is_public: this.activeFilters.is_public,
                 user_rating_min: this.activeFilters.user_rating_min ? this.activeFilters.user_rating_min : null,
@@ -581,6 +584,7 @@ export default {
                 on_shelf: false,
                 bar_shelf: false,
                 locked_bar_cocktails: false,
+                locked_user_cocktails: false,
                 favorites: false,
                 is_public: false,
                 tag_id: [],